+++ src/tasks/autoAuth.ts
